refactor: move extractDependency to common
